requirements.txt
fastapi
uvicorn
pillow
opencv-python
ultralytics
python-multipart
python-dotenv
websockets
gradio>=4.0.0
ultralytics
opencv-python